Move invariants to dedicated submodule
